Rename packages and class
