+++ app/src/main/AndroidManifest.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.mobiquity">
+    package="com.main.sample">
 
     <uses-permission android:name="android.permission.INTERNET"/>
     <application
-        android:name="com.MobiquityApp"
-        android:allowBackup="true"
+        android:name="com.App"
+        android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
