+++ app/src/main/java/com/main/MainActivity.kt
@@ -1,8 +1,9 @@
-package com.mobiquity
+package com.main
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.support.v7.app.AppCompatActivity
+import com.main.sample.R
 import com.ui.HomeFragment
 import com.utils.FragmentNavigation
 import dagger.android.AndroidInjection
