+++ app/src/main/java/com/utils/FragmentNavigation.kt
@@ -3,7 +3,7 @@ package com.utils
 import android.support.v4.app.Fragment
 import android.support.v4.app.FragmentManager
 import android.support.v7.app.AppCompatActivity
-import com.mobiquity.R
+import com.main.sample.R
 import javax.inject.Inject
 
 /**
